% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Cum_EIR_Fxn}
\alias{Cum_EIR_Fxn}
\title{Cumulative EIP. Add the daily EIP such that when EIP >= 1 the vector is infectious
Cum EIR is the parasite development rate}
\usage{
Cum_EIR_Fxn(temprature, parms)
}
\description{
Cumulative EIP. Add the daily EIP such that when EIP >= 1 the vector is infectious
Cum EIR is the parasite development rate
}
