% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{EIP_Fxn}
\alias{EIP_Fxn}
\title{calculate EIP}
\usage{
EIP_Fxn(temperature, parms)
}
\arguments{
\item{temperature}{numeric vector of daily temperatures (degrees C)}

\item{parms}{EIP coefficients}
}
\value{
numeric vector, same length as \code{temperature} the corresponding EIPs for each day (TODO: what time unit?)
}
\description{
calculate EIP
}
\examples{
temps <- 10:40
eips <- EIP_Fxn(temps, c(a = 5.3, b = -0.24, k = 0.16))
plot(temps, eips, type="l")
}
