% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{update}
\alias{update}
\title{function to update mosquito age, status and EIP
inputs: temperature data, params, EIP, timestep and max period (n)}
\usage{
update(temprature, parms, eip, timestep, n)
}
\description{
function to update mosquito age, status and EIP
inputs: temperature data, params, EIP, timestep and max period (n)
}
